Modified development setup behavior.
I removed the automatic development setup in rakefile, as it does trigger installation of bundler and gems in situations where the user is not necessarily aware that installation will happen (like running rake -T). Also, I added some extra notification to setup.rb so the user is aware that external packages will be downloaded and installed in the vendor directory.

Thanks to Ana Nelson  <ana@ananelson.com> for pointing this out.