Fixed a bug in the development environment.
It turns out that the 'yard' package creates its own incompatible implementation of "File.relative_path," which is a method used in many tasks. I resolved the incompatibility by rewriting it as convenience method "rel", renaming it as "File.relative", and substituting all occurrences of 'relative_path' as 'relative' in the dev-lib directory.